Update to Layer 137

+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLAbsChat.java
@@ -1,32 +1,32 @@
 package com.github.badoualy.telegram.tl.api;
 
+import static com.github.badoualy.telegram.tl.StreamUtils.*;
+import static com.github.badoualy.telegram.tl.TLObjectUtils.*;
+
 import com.github.badoualy.telegram.tl.core.TLObject;
 
 /**
  * Abstraction level for the following constructors:
  * <ul>
- * <li>{@link TLChannel}: channel#a14dca52</li>
- * <li>{@link TLChannelForbidden}: channelForbidden#8537784f</li>
- * <li>{@link TLChat}: chat#d91cdd54</li>
- * <li>{@link TLChatEmpty}: chatEmpty#9ba2d800</li>
- * <li>{@link TLChatForbidden}: chatForbidden#7328bdb</li>
+ * <li>{@link TLChannel}: channel#8261ac61</li>
+ * <li>{@link TLChannelForbidden}: channelForbidden#17d493d5</li>
+ * <li>{@link TLChat}: chat#41cbf256</li>
+ * <li>{@link TLChatEmpty}: chatEmpty#29562865</li>
+ * <li>{@link TLChatForbidden}: chatForbidden#6592a1a7</li>
  * </ul>
  *
- * @author Yannick Badoual yann.badoual@gmail.com
- * @see <a href="http://github.com/badoualy/kotlogram">http://github.com/badoualy/kotlogram</a>
  */
 public abstract class TLAbsChat extends TLObject {
-
-    protected int id;
+    protected long id;
 
     public TLAbsChat() {
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 }
