+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLAbsChatParticipants.java
@@ -1,29 +1,29 @@
 package com.github.badoualy.telegram.tl.api;
 
+import static com.github.badoualy.telegram.tl.StreamUtils.*;
+import static com.github.badoualy.telegram.tl.TLObjectUtils.*;
+
 import com.github.badoualy.telegram.tl.core.TLObject;
 
 /**
  * Abstraction level for the following constructors:
  * <ul>
- * <li>{@link TLChatParticipants}: chatParticipants#3f460fed</li>
- * <li>{@link TLChatParticipantsForbidden}: chatParticipantsForbidden#fc900c2b</li>
+ * <li>{@link TLChatParticipants}: chatParticipants#3cbc93f8</li>
+ * <li>{@link TLChatParticipantsForbidden}: chatParticipantsForbidden#8763d3e1</li>
  * </ul>
  *
- * @author Yannick Badoual yann.badoual@gmail.com
- * @see <a href="http://github.com/badoualy/kotlogram">http://github.com/badoualy/kotlogram</a>
  */
 public abstract class TLAbsChatParticipants extends TLObject {
-
-    protected int chatId;
+    protected long chatId;
 
     public TLAbsChatParticipants() {
     }
 
-    public int getChatId() {
+    public long getChatId() {
         return chatId;
     }
 
-    public void setChatId(int chatId) {
+    public void setChatId(long chatId) {
         this.chatId = chatId;
     }
 }
